Make inetutils-telnet and inetutils-telnetd the default implementations
The telnet and telnetd packages used to provide the netkit
implementations of these commands. But the netkit project has long been
dead upstream and sees no new releases. As long as we provide telnet
support in Debian, we should be using a project maintained upstream.

Through the use of these generic names and for telnet its standard
Priority, the netkit implementation are in effect the current Debian
defaults.

The plan has been to rename the netkit implementations to have a netkit-
binary package prefix. And then add transitional dummy packages in
inetutils to take over those, and pull in the inetutils implementations.
While cleaning up any netkit remnants during the upgrade.

This means that if people still want to use the netkit implementations
they can simply install the other packages instead, but the default
experience for users will improve.

Closes: #987729