Add systemd-sysv as an alternative dependency for a syslog provider
The initial proposal was to entirely drop the syslog dependency, but
as both portability and non-default alternatives are important, the
original proposal did not look technically sound to properly represent
what really goes on with logging, and given its tone and how it was
rushed, it seemed like spending time thinking or proposing alternatives
would be a waste of time and energy. So this seems like the best low
effort change.

Closes: #1072022, #1072023, #1072024, #1072025