Enable libsystemd support for utmp on systems running systemd
Given recent dropping of utmp support on systems due to 64-bit concerns,
the utmp files might no longer contain the necessary data for syslogd
and talkd to work as expected. Enable the gnulib libsystemd support
which gets tried and then falls back to use the traditional utmp file
to restore that functionality on systemd systems.